added animation component of login page
