+++ src/components/CanvasLogin.jsx
@@ -1,0 +1,376 @@
+import React, { useEffect, useRef } from 'react';
+
+function CanvasLogin() {
+    const canvasRef = useRef(null);
+    const animationFrameRef = useRef();
+    const eyePositionsRef = useRef({
+        left: { outer: { x: 275, y: 230 }, inner: { x: 275, y: 230 } },
+        left1: { outer: { x: 210, y: 105 }, inner: { x: 210, y: 105 } },
+        right1: { outer: { x: 240, y: 105 }, inner: { x: 240, y: 105 } },
+        right: { outer: { x: 295, y: 230 }, inner: { x: 295, y: 230 } },
+        blackLeft: { outer: { x: 170, y: 410 } },
+        blackRight: { outer: { x: 210, y: 410 } },
+        singleBlack: { outer: { x: 310, y: 320 } },
+        // Add initial position for semi-circle
+        semiCircle: { x: 190, y: 430 },
+        curvedRectangle: { x: 225, y: 125 }
+    });
+
+    useEffect(() => {
+        const canvas = canvasRef.current;
+        const container = document.querySelector('.canvas-container');
+        const c = canvas.getContext('2d');
+
+        canvas.width = container.offsetWidth;
+        canvas.height = container.offsetHeight;
+
+        const eyes = {
+            left: {
+                baseX: 275,
+                baseY: 230,
+                outerRadius: 7,
+                innerRadius: 4
+            },
+            left1: {
+                baseX: 210,
+                baseY: 105,
+                outerRadius: 7,
+                innerRadius: 4
+            },
+            right1: {
+                baseX: 240,
+                baseY: 105,
+                outerRadius: 7,
+                innerRadius: 4
+            },
+            right: {
+                baseX: 295,
+                baseY: 230,
+                outerRadius: 7,
+                innerRadius: 4
+            },
+            blackLeft: {
+                baseX: 170,
+                baseY: 410,
+                outerRadius: 6
+            },
+            blackRight: {
+                baseX: 210,
+                baseY: 410,
+                outerRadius: 6
+            },
+            singleBlack: {
+                baseX: 320,
+                baseY: 310,
+                outerRadius: 4
+            },
+            // Add base position for semi-circle
+            semiCircle: {
+                baseX: 190,
+                baseY: 430,
+                radius: 15
+            },
+            curvedRectangle: {
+                baseX: 225,
+                baseY: 125,
+                radius: 20
+            }
+        };
+
+        function calculateCurvedRectanglePosition(mouseX, mouseY, base) { }
+
+        function drawCurvedRectangle(position) {
+            //  Set drawing style
+            c.strokeStyle = '#000000';
+            c.lineWidth = 10;
+
+            // Begin drawing path
+            c.beginPath();
+
+            // Draw curved line using arc - flipped version
+            // Parameters: centerX, centerY, radius, startAngle, endAngle
+            c.arc(
+                position.x,  // x-coordinate of the center
+                position.y,  // y-coordinate of the center (above canvas)
+                eyes.curvedRectangle.radius,  // radius
+                Math.PI * 0.25,   // start angle (in radians)
+                Math.PI * 0.75,   // end angle (in radians)
+                false  // counterclockwise
+            );
+
+            // Draw the path
+            c.stroke();
+        }
+
+        function calculateSemiCirclePosition(mouseX, mouseY, base) {
+            const dx = mouseX - base.baseX;
+            const dy = mouseY - base.baseY;
+            const distance = Math.sqrt(dx * dx + dy * dy);
+            const maxMove = 10; // Maximum movement distance
+
+            let moveX = (dx / distance) * Math.min(distance * 0.1, maxMove) * 1.5;
+            let moveY = (dy / distance) * Math.min(distance * 0.1, maxMove) * 1.5;
+
+            return {
+                x: base.baseX + moveX,
+                y: base.baseY + moveY
+            };
+        }
+
+        function drawSemiCircle(position) {
+            c.beginPath();
+            c.arc(position.x, position.y, eyes.semiCircle.radius, Math.PI, 0, true);
+            c.fillStyle = 'black';
+            c.fill();
+        }
+
+        function lerp(start, end, factor) {
+            return start + (end - start) * factor;
+        }
+
+        function calculateEyePosition(mouseX, mouseY, eye, extraw, extrah) {
+            const outerOvalWidth = 20 + extraw;
+            const outerOvalHeight = 20 + extrah / 3;
+            const leftOffset = -20;
+
+            let dx = mouseX - eye.baseX;
+            let dy = mouseY - eye.baseY;
+            let angle = Math.atan2(dy, dx);
+
+            let outerX = (eye.baseX + leftOffset) + Math.cos(angle) * outerOvalWidth;
+            let outerY = eye.baseY + Math.sin(angle) * outerOvalHeight;
+
+            if (!eye.innerRadius) {
+                return {
+                    outer: { x: outerX, y: outerY }
+                };
+            }
+
+            let innerDx = mouseX - outerX;
+            let innerDy = mouseY - outerY;
+            let innerAngle = Math.atan2(innerDy, innerDx);
+            let innerDistance = 2;
+
+            let innerX = outerX + Math.cos(innerAngle) * innerDistance;
+            let innerY = outerY + Math.sin(innerAngle) * innerDistance;
+
+            return {
+                outer: { x: outerX, y: outerY },
+                inner: { x: innerX, y: innerY }
+            };
+        }
+
+        function drawEye(positions, radius, isBlackEye = false) {
+            if (isBlackEye) {
+                c.beginPath();
+                c.ellipse(positions.outer.x, positions.outer.y, radius.outerRadius,
+                    radius.outerRadius * 1.5, 0, 0, Math.PI * 2);
+                c.fillStyle = 'black';
+                c.fill();
+                return;
+            }
+
+            c.beginPath();
+            c.ellipse(positions.outer.x, positions.outer.y, radius.outerRadius,
+                radius.outerRadius * 1.5, 0, 0, Math.PI * 2);
+            c.fillStyle = 'white';
+            c.fill();
+            c.strokeStyle = 'rgba(0, 0, 0, 0.1)';
+            c.stroke();
+
+            c.beginPath();
+            c.ellipse(positions.inner.x, positions.inner.y, radius.innerRadius,
+                radius.innerRadius * 1.5, 0, 0, Math.PI * 2);
+            c.fillStyle = 'black';
+            c.fill();
+
+            const shineRadius = radius.innerRadius * 0.4;
+            c.beginPath();
+            c.arc(positions.inner.x + radius.innerRadius * 0.3,
+                positions.inner.y - radius.innerRadius * 0.3,
+                shineRadius, 0, Math.PI * 2);
+            c.fillStyle = 'rgba(255, 255, 255, 0.8)';
+            c.fill();
+        }
+
+        function animate() {
+            c.clearRect(0, 0, canvas.width, canvas.height);
+
+            if (img.complete) {
+                c.drawImage(img, 0, -90, canvas.width, canvas.height);
+            }
+
+            // Draw semi-circle first
+            drawSemiCircle(eyePositionsRef.current.semiCircle);
+
+            drawCurvedRectangle(eyePositionsRef.current.curvedRectangle);
+
+            // Draw regular eyes
+            drawEye(eyePositionsRef.current.left,
+                { outerRadius: eyes.left.outerRadius, innerRadius: eyes.left.innerRadius });
+            drawEye(eyePositionsRef.current.left1,
+                { outerRadius: eyes.left1.outerRadius, innerRadius: eyes.left1.innerRadius });
+            drawEye(eyePositionsRef.current.right1,
+                { outerRadius: eyes.right1.outerRadius, innerRadius: eyes.right1.innerRadius });
+            drawEye(eyePositionsRef.current.right,
+                { outerRadius: eyes.right.outerRadius, innerRadius: eyes.right.innerRadius });
+
+            // Draw black eyes
+            drawEye(eyePositionsRef.current.blackLeft,
+                { outerRadius: eyes.blackLeft.outerRadius }, true);
+            drawEye(eyePositionsRef.current.blackRight,
+                { outerRadius: eyes.blackRight.outerRadius }, true);
+            drawEye(eyePositionsRef.current.singleBlack,
+                { outerRadius: eyes.singleBlack.outerRadius }, true);
+
+            animationFrameRef.current = requestAnimationFrame(animate);
+        }
+
+        function handleMouseMove(e) {
+            const rect = canvas.getBoundingClientRect();
+            const mouseX = e.clientX - rect.left;
+            const mouseY = e.clientY - rect.top;
+
+            const leftPos = calculateEyePosition(mouseX, mouseY, eyes.left, 0, 0);
+            const left1Pos = calculateEyePosition(mouseX, mouseY, eyes.left1, 10, 10);
+            const right1Pos = calculateEyePosition(mouseX, mouseY, eyes.right1, 10, 10);
+            const rightPos = calculateEyePosition(mouseX, mouseY, eyes.right, 0, 0);
+            const blackLeftPos = calculateEyePosition(mouseX, mouseY, eyes.blackLeft, 15, 15);
+            const blackRightPos = calculateEyePosition(mouseX, mouseY, eyes.blackRight, 15, 15);
+            const singleBlackPos = calculateEyePosition(mouseX, mouseY, eyes.singleBlack, 5, 12);
+            const semiCirclePos = calculateSemiCirclePosition(mouseX, mouseY, eyes.semiCircle);
+            const curvedRectanglePos = calculateSemiCirclePosition(mouseX, mouseY, eyes.curvedRectangle);
+
+            const easing = 0.1;
+            eyePositionsRef.current = {
+                left: {
+                    outer: {
+                        x: lerp(eyePositionsRef.current.left.outer.x, leftPos.outer.x, easing),
+                        y: lerp(eyePositionsRef.current.left.outer.y, leftPos.outer.y, easing)
+                    },
+                    inner: {
+                        x: lerp(eyePositionsRef.current.left.inner.x, leftPos.inner.x, easing),
+                        y: lerp(eyePositionsRef.current.left.inner.y, leftPos.inner.y, easing)
+                    }
+                },
+                left1: {
+                    outer: {
+                        x: lerp(eyePositionsRef.current.left1.outer.x, left1Pos.outer.x, easing),
+                        y: lerp(eyePositionsRef.current.left1.outer.y, left1Pos.outer.y, easing)
+                    },
+                    inner: {
+                        x: lerp(eyePositionsRef.current.left1.inner.x, left1Pos.inner.x, easing),
+                        y: lerp(eyePositionsRef.current.left1.inner.y, left1Pos.inner.y, easing)
+                    }
+                },
+                right1: {
+                    outer: {
+                        x: lerp(eyePositionsRef.current.right1.outer.x, right1Pos.outer.x, easing),
+                        y: lerp(eyePositionsRef.current.right1.outer.y, right1Pos.outer.y, easing)
+                    },
+                    inner: {
+                        x: lerp(eyePositionsRef.current.right1.inner.x, right1Pos.inner.x, easing),
+                        y: lerp(eyePositionsRef.current.right1.inner.y, right1Pos.inner.y, easing)
+                    }
+                },
+                right: {
+                    outer: {
+                        x: lerp(eyePositionsRef.current.right.outer.x, rightPos.outer.x, easing),
+                        y: lerp(eyePositionsRef.current.right.outer.y, rightPos.outer.y, easing)
+                    },
+                    inner: {
+                        x: lerp(eyePositionsRef.current.right.inner.x, rightPos.inner.x, easing),
+                        y: lerp(eyePositionsRef.current.right.inner.y, rightPos.inner.y, easing)
+                    }
+                },
+                blackLeft: {
+                    outer: {
+                        x: lerp(eyePositionsRef.current.blackLeft.outer.x, blackLeftPos.outer.x, easing),
+                        y: lerp(eyePositionsRef.current.blackLeft.outer.y, blackLeftPos.outer.y, easing)
+                    }
+                },
+                blackRight: {
+                    outer: {
+                        x: lerp(eyePositionsRef.current.blackRight.outer.x, blackRightPos.outer.x, easing),
+                        y: lerp(eyePositionsRef.current.blackRight.outer.y, blackRightPos.outer.y, easing)
+                    }
+                },
+                singleBlack: {
+                    outer: {
+                        x: lerp(eyePositionsRef.current.singleBlack.outer.x, singleBlackPos.outer.x, easing),
+                        y: lerp(eyePositionsRef.current.singleBlack.outer.y, singleBlackPos.outer.y, easing)
+                    }
+                },
+                semiCircle: {
+                    x: lerp(eyePositionsRef.current.semiCircle.x, semiCirclePos.x, easing),
+                    y: lerp(eyePositionsRef.current.semiCircle.y, semiCirclePos.y, easing)
+                },
+                curvedRectangle: {
+                    x: lerp(eyePositionsRef.current.curvedRectangle.x, curvedRectanglePos.x, easing),
+                    y: lerp(eyePositionsRef.current.curvedRectangle.y, curvedRectanglePos.y, easing)
+                }
+            };
+        }
+
+        const img = new Image();
+        img.src = 'https://cdn.dribbble.com/userupload/8432950/file/original-0c14504bd291054d76548cb015dff89a.png?resize=1024x768&vertical=center';
+        img.onload = animate;
+
+        window.addEventListener('mousemove', handleMouseMove);
+
+        return () => {
+            window.removeEventListener('mousemove', handleMouseMove);
+            if (animationFrameRef.current) {
+                cancelAnimationFrame(animationFrameRef.current);
+            }
+        };
+    }, []);
+
+    return (
+        <canvas ref={canvasRef} className="max-w-full"></canvas>
+    )
+}
+
+export default CanvasLogin
+
+// import React, { useEffect, useRef } from "react";
+
+// const CurvedRectangle = () => {
+//     const canvasRef = useRef(null);
+
+//     useEffect(() => {
+//         const canvas = canvasRef.current;
+//         const ctx = canvas.getContext('2d');
+
+//         if (ctx) {
+//             // Clear canvas
+//             ctx.clearRect(0, 0, canvas.width, canvas.height);
+
+//             // Set drawing style
+//             ctx.strokeStyle = '#FF0000';
+//             ctx.lineWidth = 20;
+
+//             // Begin drawing path
+//             ctx.beginPath();
+
+//             // Draw curved line using arc - flipped version
+//             // Parameters: centerX, centerY, radius, startAngle, endAngle
+//             ctx.arc(
+//                 canvas.width / 2,  // x-coordinate of the center
+//                 -canvas.height,    // y-coordinate of the center (above canvas)
+//                 canvas.height * 1.8,  // radius
+//                 Math.PI * 0.15,   // start angle (in radians)
+//                 Math.PI * 0.85,   // end angle (in radians)
+//                 false  // counterclockwise
+//             );
+
+//             // Draw the path
+//             ctx.stroke();
+//         }
+//     }, []);
+
+//     return <canvas ref={canvasRef} width={300} height={250} style={{ display: "block", margin: "0 auto" }}></canvas>;
+// };
+
+// export default CurvedRectangle;
+
